backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Werkzeug
psycopg2-binary
